vector_pipeline/pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-openai",
    "langchain-text-splitters",
    "qdrant-client",
    "datasets",
    "pymongo",
    "python-dotenv",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "ollama>=0.6.1",
]

[dependency-groups]
dev = [
    "black>=24.0.0",
    "pylint>=3.0.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
ingest = "main:main"

[tool.pylint.messages_control]
disable = [
    "C0114", # missing-module-docstring
    "C0116", # missing-function-docstring
    "W4901", # deprecated-module (chunk.chunk conflicts with stdlib)
    "C0411", # wrong-import-order (caused by chunk.chunk stdlib conflict)
    "E1111", # assignment-from-no-return (stubs)
    "W0613", # unused-argument (stubs)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
